Try and optimize layout
