+++ src/app/dashboard/page.tsx
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation"
 
 import { auth } from "@/auth"
 import Dashboard from "@/components/dashboard"
-import { getUserByFarcasterId } from "@/db/users"
+import { getUserById } from "@/db/users"
 import { getApplications, getProjects } from "@/lib/actions/projects"
 
 export default async function Page() {
@@ -12,15 +12,15 @@ export default async function Page() {
     redirect("/")
   }
 
-  const user = await getUserByFarcasterId(session.user.farcasterId)
+  const [user, projects, applications] = await Promise.all([
+    getUserById(session.user.id),
+    getProjects(session.user.id),
+    getApplications(session.user.id),
+  ])
+
   if (!user) {
     redirect("/")
   }
-
-  const [projects, applications] = await Promise.all([
-    getProjects(session.user.farcasterId),
-    getApplications(session.user.farcasterId),
-  ])
 
   return (
     <main className="flex flex-col flex-1 h-full items-center bg-secondary pb-12">
